+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `Router::run()`
 - Route extensions are only used in `GET` and `HEAD` requests
 - Allow using a different Router in the Controller
+- `Router::requestRoot()` receive headers and request payload
 
 ### Deprecated